+++ bidsprep/_exceptions.py
@@ -3,19 +3,19 @@
 from typing import Literal, Optional
 
 
-class SliceDimensionError(Exception):
+class SliceAxisError(Exception):
     """
-    Incorrect slice dimension.
+    Incorrect slice axis.
 
     Raised when the number of slices does not match "slice_end" plus one.
 
     Parameters
     ----------
-    slice_dim: :obj:`Literal["x", "y", "z"]`
+    slice_axis: :obj:`Literal["x", "y", "z"]`
         The specified slice dimension.
 
     n_slices: :obj:`int`
-        The number of slices from the specified ``slice_dim``.
+        The number of slices from the specified ``slice_axis``.
 
     slice_end: :obj:`int`
         The number of slices specified by "slice_end" in the NIfTI header.
@@ -26,15 +26,15 @@ class SliceDimensionError(Exception):
 
     def __init__(
         self,
-        incorrect_slice_dim: Literal["x", "y", "z"],
+        slice_axis: Literal["x", "y", "z"],
         n_slices: int,
         slice_end: int,
         message: Optional[str] = None,
     ):
         if not message:
             self.message = (
-                "Incorrect slice dimension. Number of slices for "
-                f"{incorrect_slice_dim} dimension is {n_slices} but "
+                "Incorrect slice axis. Number of slices for "
+                f"{slice_axis} dimension is {n_slices} but "
                 f"'slice_end' in NIfTI header is {slice_end}."
             )
         else: